api/pyproject.toml
[tool.poetry]
name = "opsway-api"
version = "0.1.0"
description = "Opsway CI/CD Platform — API Service"
authors = ["Opsway Team"]
packages = [{include = "app"}]
[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.0"
pydantic = {extras = ["email"], version = "^2.0.0"}
pydantic-settings = "^2.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
httpx = "^0.27.0"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.0.0"
docker = "^7.1.0"
gitpython = "^3.1.43"
minio = "^7.2.0"
python-multipart = "^0.0.9"
aiofiles = "^24.1.0"
anyio = "^4.4.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
